refactor routines
