+++ src/DiscordBot.ts
@@ -1,8 +1,7 @@
-import { Channel, Client, Message, RichEmbed, TextChannel } from 'discord.js';
+import { Client, Message } from 'discord.js';
 
 import { CommandFactory } from './commands/CommandFactory';
-import { buildEmbed } from './utils/buildEmbed';
-import { SubredditFetcher } from './reddit/SubredditFetcher';
+import { startFetchAndPostRoutine } from './routines/fetchAndPost.routine';
 
 export class DiscordBot {
   private prefix: string = 'z!';
@@ -32,7 +31,8 @@ export class DiscordBot {
       console.log('Discord Bot connected');
       await this.client.user.setActivity('with sparkles');
 
-      // await this.startFetchingRoutine();
+      //* routines
+      await startFetchAndPostRoutine(30, this.client);
     });
   };
 
@@ -46,29 +46,5 @@ export class DiscordBot {
       const command = this.commandFactory.createCommand(message);
       command.execute();
     });
-  };
-
-  private startFetchingRoutine = async (): Promise<void> => {
-    const delaySeconds = 300;
-    
-    const chan = this.client.channels.find(
-      channel => channel.id === '675271307696406545'
-    );
-
-    const fetchAndPost = async () => {
-      console.log('Started fetching subreddit posts');
-      const fetcher = new SubredditFetcher('zoemains');
-      const posts = await fetcher.getLatestPostsSince(delaySeconds);
-      posts.forEach(async post => {
-        console.log(`> Posting: ${post.title}`);
-        const embed: RichEmbed = buildEmbed(post);
-        await (chan as TextChannel).send(embed);
-      });
-      console.log('Done fetching subreddit posts');
-    };
-
-    await fetchAndPost();
-
-    setInterval(fetchAndPost, delaySeconds * 1000);
   };
 }
