+++ src/commands/Ping.command.ts
