--- src/functions/FetchAndPost.ts
@@ -1,3 +1,0 @@
-export class FetchAndPost {
-  
-}
